+++ final.RayTracing/code/src/object3d.hpp
@@ -20,9 +20,9 @@ public:
     }
 
     // Intersect Ray with this object. If hit, store information in hit structure.
-    virtual bool intersect(const Ray &r, Hit &h, float tmin) = 0;
+    virtual bool intersect(const Ray &r, Hit &h, double tmin) = 0;
 
-    inline float abs_f(float x) { return (x<0 ? -x : x);}
+    inline double abs_f(double x) { return (x<0 ? -x : x);}
 
 };
 
@@ -32,7 +32,7 @@ class Plane : public Object3D {
 public:
     Vec3 p;
     Vec3 normal;
-    float d;
+    double d;
 
     Plane(): Object3D() {}
 
@@ -43,15 +43,15 @@ public:
 
     ~Plane() override = default;
 
-    bool intersect(const Ray &r, Hit &h, float tmin) override {
+    bool intersect(const Ray &r, Hit &h, double tmin) override {
         // ray r is parallel with the plane
-        float m = r.dir.dot(normal);
+        double m = r.dir.dot(normal);
         if (abs_f(m) < 1e-9)
             return false;
 
         // not parallel
-        float s = d - normal.dot(r.origin);
-        float t = s / m; 
+        double s = d - normal.dot(r.origin);
+        double t = s / m; 
         if (t > tmin) {
             h.set(t, material, normal);
             return true;
@@ -59,9 +59,9 @@ public:
         return false;
     }
 
-    // float solve(const Ray &r) {
-    //     float s = d - normal.dot(r.origin);
-    //     float t = s / normal.dot(r.dir);
+    // double solve(const Ray &r) {
+    //     double s = d - normal.dot(r.origin);
+    //     double t = s / normal.dot(r.dir);
     //     return t;
     // }
 };
@@ -69,29 +69,29 @@ public:
 class Sphere : public Object3D {
 public:
     Vec3 center;
-    float radius;
+    double radius;
 
     // unit ball at the center
     Sphere() : center(), radius(1.) {}
 
-    Sphere(const Vec3 &_center, float _radius, Material *_material) :
+    Sphere(const Vec3 &_center, double _radius, Material *_material) :
         Object3D(_material), center(_center), radius(_radius) {}
 
     ~Sphere() override = default;
 
-    bool intersect(const Ray &r, Hit &h, float tmin) override {
+    bool intersect(const Ray &r, Hit &h, double tmin) override {
         // Calc dist from center to ray r
         Vec3 originToCent = center - r.origin;
-        float d1 = abs_f(r.dir.normalized().dot(originToCent));
-        float d2 = originToCent.len2() - d1 * d1; // center to ray r
+        double d1 = abs_f(r.dir.normalized().dot(originToCent));
+        double d2 = originToCent.len2() - d1 * d1; // center to ray r
         // Cals intersection
-        float interHalfLen = sqrt(radius*radius - d2);
-        float t = d1 - interHalfLen;
+        double interHalfLen = sqrt(radius*radius - d2);
+        double t = d1 - interHalfLen;
         if (t < tmin) return false;
         else {
             Vec3 normal = r.pointAtParameter(t) - center;
-            float u = 0.5 + atan2(normal.z, normal.x) / (2. * M_PI);
-            float v = 0.5 - asin(normal.y) / M_PI;
+            double u = 0.5 + atan2(normal.z, normal.x) / (2. * M_PI);
+            double v = 0.5 - asin(normal.y) / M_PI;
             h.set(t, material, normal.normalized(), Vec3(u, v));
             return true;
         }
@@ -120,7 +120,7 @@ public:
 
     ~Transform() {}
 
-    virtual bool intersect(const Ray &r, Hit &h, float tmin) {
+    virtual bool intersect(const Ray &r, Hit &h, double tmin) {
         Vec3 trSource = transformPoint(transform, r.origin);
         Vec3 trDirection = transformDirection(transform, r.dir);
         Ray tr(trSource, trDirection);
@@ -149,7 +149,7 @@ public:
 	Triangle( const Vec3& a, const Vec3& b, const Vec3& c, Material* m) :
         Plane(b-a.cross(c-a), a, m), vertices{a, b, c} {}
 
-	bool intersect( const Ray& ray,  Hit& hit , float tmin) override {
+	bool intersect( const Ray& ray,  Hit& hit , double tmin) override {
 		// solve the intersection between ray and the triangle plane
         Hit h_tmp;
         if (Plane::intersect(ray, h_tmp, tmin)) {
@@ -161,9 +161,9 @@ public:
             Vec3 v0 = vertices[2] - vertices[0];
             Vec3 v1 = vertices[1] - vertices[0];
             Vec3 v2 = p - vertices[0];
-            float denominator = v0.len2()*v1.len2() - v0.dot(v1)*v1.dot(v0);
-            float u = (v1.len2()*v2.dot(v0) - v1.dot(v0)*v2.dot(v1)) / denominator;
-            float v = (v0.len2()*v2.dot(v1) - v0.dot(v1)*v2.dot(v0)) / denominator;
+            double denominator = v0.len2()*v1.len2() - v0.dot(v1)*v1.dot(v0);
+            double u = (v1.len2()*v2.dot(v0) - v1.dot(v0)*v2.dot(v1)) / denominator;
+            double v = (v0.len2()*v2.dot(v1) - v0.dot(v1)*v2.dot(v0)) / denominator;
             if (good(u) && good(v) && good(u+v)) {
                 hit = h_tmp;
                 return true;
@@ -172,13 +172,13 @@ public:
         return false;
 	}
 	
-	inline bool good(float x) { return (0 <= x && x <= 1); }
+	inline bool good(double x) { return (0 <= x && x <= 1); }
 };
 
 class Rectangle: public Plane {
 public:
     Vec3 u, v;
-    float u_len_inv, v_len_inv;
+    double u_len_inv, v_len_inv;
 
     Rectangle() {}
 
@@ -187,12 +187,12 @@ public:
         v(v_.normalized()), u_len_inv(1. / u_.len()),
         v_len_inv(1. / v_.len()) {}
     
-    bool intersect(const Ray& r, Hit& h, float tmin) override {
+    bool intersect(const Ray& r, Hit& h, double tmin) override {
         Hit h_tmp;
         if (Plane::intersect(r, h_tmp, tmin)) {
             Vec3 p_to_x = r.pointAtParameter(h_tmp.t) - p; // p -> point of intersection
-            float du = p_to_x.dot(u) * u_len_inv;
-            float dv = p_to_x.dot(v) * v_len_inv;
+            double du = p_to_x.dot(u) * u_len_inv;
+            double dv = p_to_x.dot(v) * v_len_inv;
             if (tmin < du && du < 1-tmin && tmin < dv && dv < 1-tmin) {
                 h.set(h_tmp.t, h_tmp.material, h_tmp.normal, Vec3(du, dv));
                 return true;
@@ -205,10 +205,10 @@ public:
 
 class Circle: public Plane {
 public:
-    float r;
+    double r;
     Vec3 u, v;
 
-    Circle(const Vec3& c, float r_, const Vec3& n, Material *m): Plane(n, c, m), r(r_) {
+    Circle(const Vec3& c, double r_, const Vec3& n, Material *m): Plane(n, c, m), r(r_) {
         if (n.x < 1.0) {
             u.x = n.z;
             u.y = 0.;
@@ -222,11 +222,11 @@ public:
         v = n.cross(u).normalized() * r;
     }
 
-    bool intersect(const Ray& ray, Hit& h, float tmin) override {
+    bool intersect(const Ray& ray, Hit& h, double tmin) override {
         Hit h_tmp;
         if (Plane::intersect(ray, h_tmp, tmin)) {
             Vec3 x = ray.pointAtParameter(h_tmp.t);
-            float dist = (x - p).len();
+            double dist = (x - p).len();
             if (dist < r) {
                 h = h_tmp;
                 return true;
